pyproject.toml
[tool.poetry]
name = "uviversity-app"
version = "0.1.0"
description = ""
authors = ["iva <ivalulyanov1@gmail.com>"]
license = "LICENSE"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.114.0"
uvicorn = "^0.30.6"
starlette = "^0.38.5"
python-dotenv = "^1.0.1"
jinja2 = "^3.1.4"
python-multipart = "^0.0.9"
sqlalchemy = "^2.0.34"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
